Extract Base UI config to a separate directory
