+++ packages/api-docs-builder/cli.ts
@@ -2,13 +2,9 @@ import path from 'path';
 import * as yargs from 'yargs';
 import { ProjectSettings, buildApi } from './buildApi';
 import findApiPages from './utils/findApiPages';
+import baseUiProjectSettings from './projects/baseUi/projectSettings';
 import muiSystemSettings from './projects/muiSystem/projectSettings';
-import {
-  getBaseComponentInfo,
-  getBaseHookInfo,
-  getJoyComponentInfo,
-  getMaterialComponentInfo,
-} from './buildApiUtils';
+import { getJoyComponentInfo, getMaterialComponentInfo } from './buildApiUtils';
 
 type CommandOptions = { grep?: string };
 
@@ -21,15 +17,7 @@ const projectSettings: ProjectSettings[] = [
     getApiPages: () => findApiPages('docs/pages/material-ui/api'),
     getComponentInfo: getMaterialComponentInfo,
   },
-  {
-    output: {
-      apiManifestPath: path.join(process.cwd(), 'docs/data/base/pagesApi.js'),
-    },
-    typeScriptProjects: ['base'],
-    getApiPages: () => findApiPages('docs/pages/base-ui/api'),
-    getComponentInfo: getBaseComponentInfo,
-    getHookInfo: getBaseHookInfo,
-  },
+  baseUiProjectSettings,
   {
     output: {
       apiManifestPath: path.join(process.cwd(), 'docs/data/joy/pagesApi.js'),
@@ -49,7 +37,7 @@ async function run(argv: yargs.ArgumentsCamelCase<CommandOptions>) {
 yargs
   .command({
     command: '$0',
-    describe: 'formats codebase',
+    describe: 'Generates API documentation for the MUI packages.',
     builder: (command) => {
       return command.option('grep', {
         description:
