+++ packages/api-docs-builder/projects/baseUi/getBaseUiHookInfo.ts
@@ -1,0 +1,72 @@
+import fs from 'fs';
+import path from 'path';
+import kebabCase from 'lodash/kebabCase';
+import { getHeaders, getTitle } from '@mui/markdown';
+import {
+  ComponentInfo,
+  HookInfo,
+  extractPackageFile,
+  fixPathname,
+  getApiPath,
+  parseFile,
+} from '../../buildApiUtils';
+import findPagesMarkdown from '../../utils/findPagesMarkdown';
+import migratedBaseComponents from './migratedBaseComponents';
+
+export default function getBaseHookInfo(filename: string): HookInfo {
+  const { name } = extractPackageFile(filename);
+  let srcInfo: null | ReturnType<ComponentInfo['readFile']> = null;
+  if (!name) {
+    throw new Error(`Could not find the hook name from: ${filename}`);
+  }
+
+  const allMarkdowns = findPagesMarkdown()
+    .filter((markdown) => {
+      if (migratedBaseComponents.some((component) => filename.includes(component))) {
+        return markdown.filename.match(/[\\/]data[\\/]base[\\/]/);
+      }
+      return true;
+    })
+    .map((markdown) => {
+      const markdownContent = fs.readFileSync(markdown.filename, 'utf8');
+      const markdownHeaders = getHeaders(markdownContent) as any;
+
+      return {
+        ...markdown,
+        markdownContent,
+        hooks: markdownHeaders.hooks as string[],
+      };
+    });
+
+  const demos = findBaseHooksDemos(name, allMarkdowns);
+  const apiPath = getApiPath(demos, name);
+
+  const result = {
+    filename,
+    name,
+    apiPathname: apiPath ?? `/base-ui/api/${kebabCase(name)}/`,
+    apiPagesDirectory: path.join(process.cwd(), `docs/pages/base-ui/api`),
+    readFile: () => {
+      srcInfo = parseFile(filename);
+      return srcInfo;
+    },
+    getDemos: () => demos,
+  };
+  return result;
+}
+
+function findBaseHooksDemos(
+  hookName: string,
+  pagesMarkdown: ReadonlyArray<{
+    pathname: string;
+    hooks: readonly string[];
+    markdownContent: string;
+  }>,
+) {
+  return pagesMarkdown
+    .filter((page) => page.hooks && page.hooks.includes(hookName))
+    .map((page) => ({
+      demoPageTitle: getTitle(page.markdownContent),
+      demoPathname: `${fixPathname(page.pathname)}#hook${page.hooks?.length > 1 ? 's' : ''}`,
+    }));
+}
