+++ packages/api-docs-builder/projects/baseUi/migratedBaseComponents.ts
@@ -1,0 +1,23 @@
+export default [
+  'Badge',
+  'Button',
+  'ClickAwayListener',
+  'FocusTrap',
+  'Input',
+  'MenuItem',
+  'Menu',
+  'Modal',
+  'NoSsr',
+  'OptionGroup',
+  'Option',
+  'Popper',
+  'Portal',
+  'Select',
+  'Slider',
+  'Switch',
+  'TablePagination',
+  'TabPanel',
+  'TabsList',
+  'Tabs',
+  'Tab',
+];
